gate

+++ app/Services/Admin/BannerService.php
@@ -3,6 +3,7 @@ namespace App\Services\Admin;
 
 use App\Traits\WebResponseTrait;
 use App\Models\Banner;
+use Illuminate\Support\Facades\Gate;
 
 class BannerService
 {
@@ -10,14 +11,18 @@ class BannerService
 
     public function index($request)
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         $builder = Banner::where(function ($query) use ($request) {
             if ($request->name) $query->where('name', 'like', '%'.$request->name.'%');
         });
 
         $data = $builder->orderBy('id', 'desc')
-                        ->paginate(10);
+                        ->get();
 
-        $data->appends(request()->query());
+        // $data->appends(request()->query());
 
         return view(
             'admin::banner.index',
@@ -27,11 +32,19 @@ class BannerService
 
     public function create()
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         return view('admin::banner.create');
     }
 
     public function store($request)
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         $data = request()->all();
         $data['image'] = $request->file('image')->store('banner', 'public');
         $data['active'] = STATUS_ACCOUNT_CUSTOMER_NOT_ACTIVE;
@@ -42,6 +55,10 @@ class BannerService
 
     public function show($id)
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         $data = Banner::find($id);
         return view(
             'admin::banner.edit',
@@ -53,6 +70,10 @@ class BannerService
 
     public function update($request, $id)
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         $data=$request->all();
         $banner = Banner::find($id);
         if(empty($banner)) {
@@ -69,10 +90,19 @@ class BannerService
     }
     public function delete($id)
     {
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         Banner::where('id', $id)->delete();
         return $this->returnSuccessWithRoute('admin.banner.index', __('messages.data_delete_success'));
     }
     public function listSoftDelete(){
+
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         $data = Banner::onlyTrashed()->get();
         return view(
             'admin::banner.listSoftDelete',
@@ -81,6 +111,11 @@ class BannerService
     }
     public function restore($id)
     {
+
+        if (! Gate::allows('Quản trị viên')) {
+            return abort(401);
+        }
+
         Banner::withTrashed()->where('id', $id)->restore();
         return $this->returnSuccessWithRoute('admin.banner.listSoftDelete', __('messages.data_create_success'));
     }
